PyCyberSuite/requirements.txt
tkinter
cryptography
requests
python-nmap
schedule